+++ ComponentLib/PowerElectronics/Resistor/Resistor.hpp
@@ -5,7 +5,7 @@
 
 #include <ModelEvaluatorImpl.hpp>
 #include <PowerSystemData.hpp>
-#include <ComponentLib/PowerElectronicsComponents/CircuitComponent.hpp>
+#include <ComponentLib/PowerElectronics/CircuitComponent.hpp>
 
 
 namespace ModelLib
